Major graphics overhaul and other stuff
I did a ton of stuff. I made and added a background because the plain blue was ugly. I changed the bricks to grass because we were going to start with the grassland world. The blocks i added were barriers (invisible solid blocks), 4 types of grass blocks, dirt corner blocks, and more ( i think i dont remember). I messed up by changing the origin position of blocks but i soon realized my mistake and fixed it. The parkour may be slightly different due to this. Gamer, if you're reading this, i wanted to make the clouds on the background to move. Talk to me about this. Anyways yeah.