Added dirt textures + BIG ISSUE
So I did a lot of stuff like name the game butt bashers and added our company so anti-viruses no longer say our game is a virus now. I added an icon, added a splash screen, and I did all of the miscellaneous dirt textures like the corners and stuff. I added more to the level and added decorations. (also an obvious easter egg). So yeah anyways that big problem. So gamer named the game project "New Game", and game maker uses the project name as the game name so when I install the game on my computer, its not called butt bashers. It just ends up being called New Game, so we wanna fix that really soon.