Added some stuff - read desc.
So I added some background music. I know gamer, you're gonna be like "cHiP, yOuR cOdInG sUcKs", so go ahead and fix it then if you wanna waste your time. It works, it might not be the most efficient coding possible to ever exist in game maker, but it works. "If it ain't broke, don't fix it" - Bert Lance. I made it so that you can use arrow keys in addition to WASD, and I did it so that its totally adaptable, so if you change where WASD is being used, it will still work perfectly. Also, i got rid of the "fucking idiot" sign, because this game is strictly PG, NIGGER. I'm sorry Gamer Haxer Fam that I did coding. "What we've got here is failure to communicate." - Cool Hand Luke. ----- "Well, nobody's perfect." - Some Like it Hot ----- “Two things are infinite: the universe and human stupidity; and I'm not sure about the universe.” - Albert Einstein ---- Now time for some hitler quotes.

If you tell a big enough lie and tell it frequently enough, it will be believed.

Make the lie big, make it simple, keep saying it, and eventually they will believe it.

He alone, who owns the youth, gains the future.

Those who want to live, let them fight, and those who do not want to fight in this world of eternal struggle do not deserve to live.

Demoralize the enemy from within by surprise, terror, sabotage, assassination. This is the war of the future.

The great masses of the people will more easily fall victims to a big lie than to a small one.

Words build bridges into unexplored regions.

As a Christian I have no duty to allow myself to be cheated, but I have the duty to be a fighter for truth and justice.

The victor will never be asked if he told the truth.

All propaganda has to be popular and has to accommodate itself to the comprehension of the least intelligent of those whom it seeks to reach.

- Adolf Hitler

 "SHUT THE FUCK UP" - Chip ---- "GAMER GO KILL YOURSELF" - Chip.

And that was a history of famous quotes.

im sorry