Added and tested sand stuff
I noticed the game got REEEEEAAAAALY laggy in the sand level. I think it might be because of my game maker so im commiting this and restarting game maker. If it's not, i'm gonna try to look at the console while im in the level and see if theres anything causing the game to slow down. Anyways, just added some stuff so yeh.