Added checkpoint text and more
So i added text that spawns above the checkpoint and it also makes a noise now btw. I kept trying to make it sparkle when you touch a coin, but it didnt work. So try to fix that gamer. Also, my level i made is soooooouuuper laggy, so fix that too gamer. Anyways, yeah. Chippy Dippy outie.

ps: the eagle was for a joke thing but it didnt work lol