Added lots of little stuff (read desc)
Heres a list of changes:

- New level map
- Heart powerup noise
- Dev mode
- Level unlocking
- Hand cursor animation
- Hand cursor click sound

To activate dev mode, theres a button in the options menu. If you click it, it should turn green and then dev mode is active. This allows for the numbers on your keyboard to be used for teleporting. I added this because I didn't want normal players to just teleport everywhere.

Like usual, maybe I forgot about something but idk. Anyways yeah.

That's pretty much it. Hope you like it. Chippy dippy outie.

PS: I want to change the desert from a dry and clay-rich desert to a more tropical desert with palm trees and stuff. I just think it looks cooler that way.